% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anygender.R
\name{any_gender}
\alias{any_gender}
\title{any_gender}
\usage{
any_gender(text)
}
\arguments{
\item{text}{a character vector}
}
\value{
A list
}
\description{
Calculates and stores occurance of gendered words
}
